app/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
gunicorn==21.2.0

psycopg2-binary==2.9.9

redis==5.0.4
celery==5.3.6

sqlalchemy==2.0.29
python-dotenv==1.0.1

pydantic==2.6.4
pydantic-settings==2.2.1
email-validator==2.1.1